refactor(orc8r): Indexer service security changes
