+++ orc8r/cloud/go/services/state/test_init/test_service_init.go
@@ -25,7 +25,8 @@ import (
 	"magma/orc8r/cloud/go/services/state"
 	"magma/orc8r/cloud/go/services/state/indexer/reindex"
 	indexer_protos "magma/orc8r/cloud/go/services/state/protos"
-	servicers "magma/orc8r/cloud/go/services/state/servicers/protected"
+	"magma/orc8r/cloud/go/services/state/servicers"
+	indexermgr_servicers "magma/orc8r/cloud/go/services/state/servicers/protected"
 	"magma/orc8r/cloud/go/sqorc"
 	"magma/orc8r/cloud/go/test_utils"
 	"magma/orc8r/lib/go/protos"
@@ -62,7 +63,7 @@ func startService(t *testing.T, db *sql.DB) (reindex.Reindexer, reindex.JobQueue
 	queue := reindex.NewSQLJobQueue(singleAttempt, db, sqorc.GetSqlBuilder())
 	require.NoError(t, queue.Initialize())
 	reindexer := reindex.NewReindexerQueue(queue, reindex.NewStore(factory))
-	indexerServicer := servicers.NewIndexerManagerServicer(reindexer, false)
+	indexerServicer := indexermgr_servicers.NewIndexerManagerServicer(reindexer, false)
 	indexer_protos.RegisterIndexerManagerServer(srv.GrpcServer, indexerServicer)
 
 	go srv.RunTest(lis)
@@ -90,7 +91,7 @@ func startSingletonService(t *testing.T, db *sql.DB) reindex.Reindexer {
 	versioner := reindex.NewVersioner(db, sqorc.GetSqlBuilder())
 	versioner.Initialize()
 	reindexer := reindex.NewReindexerSingleton(reindex.NewStore(factory), versioner)
-	indexerServicer := servicers.NewIndexerManagerServicer(reindexer, false)
+	indexerServicer := indexermgr_servicers.NewIndexerManagerServicer(reindexer, false)
 	indexer_protos.RegisterIndexerManagerServer(srv.GrpcServer, indexerServicer)
 
 	go srv.RunTest(lis)
